requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9  # For Alembic migrations
alembic==1.12.1

# Redis
redis==5.0.1

# MinIO
minio==7.2.0
aioboto3==12.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
authlib==1.2.1  # For OAuth
google-auth==2.23.4  # For Google OAuth verification
requests==2.31.0  # Required by google-auth

# Email
aiosmtplib==3.0.1
email-validator==2.1.0
jinja2==3.1.2

# Utils
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
python-dotenv==1.0.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0 